requirements.txt
flask
twilio
supabase
python-dotenv
requests
resend
groq
pytest
